+++ app/map/header/styles.tsx
@@ -1,0 +1,21 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+  header: {
+    marginTop: "10%",
+    flexDirection: "row",
+    gap: 20,
+    alignItems: "center",
+    padding: 10,
+    backgroundColor: "transparent",
+    width: "100%",
+  },
+  title: {
+    fontSize: 13,
+    fontFamily: "NotoBold",
+  },
+
+  backButton: {
+    width: 30,
+  },
+});
+export default styles;
